Add confirmation dialog to item deletion in V2 sheet

+++ module/actor/actor-sheet-v2.js
@@ -304,8 +304,29 @@ export class ActorArchmageSheetV2 extends ActorSheet {
     if (!itemId) return;
 
     // Delete the item from the actor object.
-    let item = this.actor.items.get(itemId);
-    await item.delete();
+    let del = false;
+    new Dialog({
+      title: game.i18n.localize("ARCHMAGE.CHAT.DeleteConfirm"),
+      // content: game.i18n.localize("ARCHMAGE.CHAT.NoUsesMsg"),
+      buttons: {
+        del: {
+          label: game.i18n.localize("ARCHMAGE.CHAT.Delete"),
+          callback: () => {del = true;}
+        },
+        cancel: {
+          label: game.i18n.localize("ARCHMAGE.CHAT.Cancel"),
+          callback: () => {}
+        }
+      },
+      default: 'cancel',
+      close: html => {
+        if (del) {
+          let item = this.actor.items.get(itemId);
+          // await item.delete();
+          item.delete();
+        }
+      }
+    }).render(true);
   }
 
   _editItem(event) {
